[SubtasksJS-6] Работа с константами, разбивка функции, и вынос логики в отдельную функцию для обработки ошибок файла.
