+++ SubtasksJs/Nika/SubtaskJS-6/src/js/const.js
@@ -1,0 +1,4 @@
+export const visibleErrorClassName = 'form__error_visible';
+export const formResultID = 'result';
+export const blockErrorID = 'error';
+export const blockError = document.getElementById(blockErrorID);
